requirements.txt
# Telegram Bot
aiogram>=3.15.0

# Supabase
supabase>=2.10.0

# Environment
python-dotenv==1.0.0

# Data validation
pydantic>=2.5.3

# Logging
loguru==0.7.2
